// src/components/cards/designs/types.ts
import React, { ReactNode, CSSProperties } from 'react';

export interface NewsDesignProps {
  title: string;
  image: string;
  image2?: string;
  formattedDate: ReactNode | string;
  themeColor: string;
  customLogo?: string;
  fullBrandLogo?: string;
  fullBrandLogoHeight?: number;
  brandName: string;
  website: string;
  qrValue: string;
  headlineFontSize?: number;
  headlineColor?: string;
  brandFontSize?: number;
  brandColor?: string;
  textAlign?: "left" | "center" | "right" | "justify";
  gradientStart?: string;
  gradientEnd?: string;
  cardGradientStart?: string;
  cardGradientEnd?: string;
  applyGradientToAll?: boolean;
  description?: string;
  hashtag?: string;
  showSocialIcons?: boolean;
  descriptionFontSize?: number;
  descriptionColor?: string;
  descriptionBgColor?: string;
  showDescriptionBg?: boolean;
  descriptionTextAlign?: "left" | "center" | "right" | "justify";
  descriptionOffsetX?: number;
  descriptionOffsetY?: number;
  overlayOpacity?: number;
  showGeometricShapes?: boolean;
  showDetailedNewsBox?: boolean;
  geometricShapeColor?: string;
  geometricShapeOpacity?: number;
  customDateColor?: string;
  customDateBgColor?: string;
  customDetailsTextColor?: string;
  customVisitTextColor?: string;
  customLogoTextColor?: string;
  customLogoBgColor?: string;
  customQrColor?: string;
  customSocialIconColor?: string;
  renderFormattedTitle: () => ReactNode;
  renderImage: (className: string, customImageSrc?: string) => ReactNode;
  renderBackgroundPattern: () => ReactNode;
  videoFit?: "cover" | "contain";
  videoBgColor?: string;
  FloralPattern: React.FC<{ color?: string; opacity?: number }>;
  brandStyle: CSSProperties;
  headlineStyle: CSSProperties;
}